third_party/blink/web_tests/platform/linux/external/wpt/pointerevents/pointerevent_pointercapture_in_frame_pen-expected.txt

This is a testharness.js-based test.
[FAIL] Test penpointer capture in same-origin frame: Pointer down at outer frame body and set pointer capture.
  assert_equals: Expected pointerup event, but got lostpointercapture event instead expected "pointerup" but got "lostpointercapture"
Harness: the test ran to completion.
